Task-8/requirements.txt
streamlit
transformers
torch
google-generativeai
nltk
python-dotenv
sentencepiece
accelerate